move district to own stack, modularize district match screens
