+++ src/screens/DistrictMatch/screens/index.js
@@ -1,0 +1,4 @@
+export { default as Address } from './Address';
+export { default as ZipCode } from './ZipCode';
+export { default as Success } from './Success';
+export { default as Failure } from './Failure';
